requirements.txt
fastapi==0.111.1
hypercorn==0.17.3
pydantic==2.8.2
pyppeteer==2.0.0
python-dotenv==1.0.1
Requests==2.32.3
aiohttp==3.9.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
llama-cpp-python
transformers
torch
torchvision
pillow
